requirements.txt
openai>=1.0.0
anthropic>=0.5.0
smolagents>=0.1.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
cohere>=4.32
langchain>=0.0.275
langchain-core>=0.1.1
langchain-community>=0.0.1
langchain-iris>=0.1.0
langchain-openai>=0.0.1
openai>=1.0.0
smolagents[all]>=0.1.0
anthropic>=0.5.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
spacy>=3.4.0
pdfplumber
reportlab